+++ Userland/Libraries/LibCore/ProcessStatisticsReader.h
@@ -22,12 +22,12 @@ struct ThreadStatistics {
     unsigned inode_faults;
     unsigned zero_faults;
     unsigned cow_faults;
-    unsigned unix_socket_read_bytes;
-    unsigned unix_socket_write_bytes;
-    unsigned ipv4_socket_read_bytes;
-    unsigned ipv4_socket_write_bytes;
-    unsigned file_read_bytes;
-    unsigned file_write_bytes;
+    u64 unix_socket_read_bytes;
+    u64 unix_socket_write_bytes;
+    u64 ipv4_socket_read_bytes;
+    u64 ipv4_socket_write_bytes;
+    u64 file_read_bytes;
+    u64 file_write_bytes;
     DeprecatedString state;
     u32 cpu;
     u32 priority;
